requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg==3.1.19
psycopg-binary==3.1.19
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
